turn on higher warning level in msvc for the core and libteletone and resolve warnings.
git-svn-id: http://svn.freeswitch.org/svn/freeswitch/trunk@634 d0543943-73ff-0310-b7d9-9358b9ac24b2

+++ src/include/switch_buffer.h
@@ -80,7 +80,7 @@ SWITCH_DECLARE(int) switch_buffer_freespace(switch_buffer *buffer);
  * \param buffer any buffer of type switch_buffer
  * \return int ammount of buffer curently in use
  */
-SWITCH_DECLARE(int) switch_buffer_inuse(switch_buffer *buffer);
+SWITCH_DECLARE(size_t) switch_buffer_inuse(switch_buffer *buffer);
 
 /*! \brief Read data from a switch_buffer up to the ammount of datalen if it is available.  Remove read data from buffer. 
  * \param buffer any buffer of type switch_buffer
